Refactored stack to data_structure
